Rename QueryBackendHandler to DatabaseBackendHandler (#8489)
