+++ shardingsphere-proxy/shardingsphere-proxy-backend/src/test/java/org/apache/shardingsphere/proxy/backend/text/admin/DALBackendHandlerFactoryTest.java
@@ -20,7 +20,7 @@ package org.apache.shardingsphere.proxy.backend.text.admin;
 import lombok.SneakyThrows;
 import org.apache.shardingsphere.proxy.backend.communication.jdbc.connection.BackendConnection;
 import org.apache.shardingsphere.proxy.backend.text.TextProtocolBackendHandler;
-import org.apache.shardingsphere.proxy.backend.text.query.QueryBackendHandler;
+import org.apache.shardingsphere.proxy.backend.text.database.DatabaseBackendHandler;
 import org.apache.shardingsphere.sql.parser.sql.common.statement.dal.DALStatement;
 import org.apache.shardingsphere.sql.parser.sql.common.statement.dal.SetStatement;
 import org.apache.shardingsphere.sql.parser.sql.dialect.statement.mysql.dal.MySQLShowDatabasesStatement;
@@ -81,12 +81,12 @@ public final class DALBackendHandlerFactoryTest {
     }
     
     @Test
-    public void assertUnicastBackendHandlerReturnedWhenOtherDALStatement() {
+    public void assertDatabaseBackendHandlerReturnedWhenOtherDALStatement() {
         DALStatement dalStatement = mock(DALStatement.class);
         BackendConnection backendConnection = mock(BackendConnection.class);
         TextProtocolBackendHandler textProtocolBackendHandler = DALBackendHandlerFactory.newInstance(dalStatement, "", backendConnection);
-        assertThat(textProtocolBackendHandler, instanceOf(QueryBackendHandler.class));
-        QueryBackendHandler backendHandler = (QueryBackendHandler) textProtocolBackendHandler;
+        assertThat(textProtocolBackendHandler, instanceOf(DatabaseBackendHandler.class));
+        DatabaseBackendHandler backendHandler = (DatabaseBackendHandler) textProtocolBackendHandler;
         assertFieldOfInstance(backendHandler, "sqlStatement", is(dalStatement));
         assertFieldOfInstance(backendHandler, "sql", is(""));
         assertFieldOfInstance(backendHandler, "backendConnection", is(backendConnection));
